fix(permission-table): fix unit test a. 修复单元测试
2024-04-16

+++ server/odc-service/src/main/java/com/oceanbase/odc/service/permission/table/TablePermissionService.java
@@ -429,6 +429,9 @@ public class TablePermissionService {
         }
         List<Long> permissionIds = userTablePermissionRepository.findByDatabaseIdIn(databaseIds).stream().map(
                 UserTablePermissionEntity::getId).collect(Collectors.toList());
+        if (CollectionUtils.isEmpty(permissionIds)) {
+            return;
+        }
         permissionRepository.deleteByIds(permissionIds);
         userPermissionRepository.deleteByPermissionIds(permissionIds);
     }
